Modularize user controller
Refactored user controller to modularize the code into files.
